requirements.txt
Streamlit
requests
pandas
numpy
scikit-learn
nltk
python-dotenv